requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
geographiclib==2.0
geopy==2.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
Werkzeug==3.0.1


